aura tests pass
