+++ core/client/src/client.rs
@@ -1209,6 +1209,21 @@ impl<B, E, Block, RA> BlockBody<Block> for Client<B, E, Block, RA>
 	}
 }
 
+impl<B, E, Block, RA> backend::AuxStore for Client<B, E, Block, RA>
+	where
+		B: backend::Backend<Block, Blake2Hasher>,
+		E: CallExecutor<Block, Blake2Hasher>,
+		Block: BlockT<Hash=H256>,
+{
+	/// Insert auxiliary data into key-value store.
+	fn insert_aux<'a, 'b: 'a, 'c: 'a, I: IntoIterator<Item=&'a(&'c [u8], &'c [u8])>, D: IntoIterator<Item=&'a &'b [u8]>>(&self, insert: I, delete: D) -> error::Result<()> {
+		::backend::AuxStore::insert_aux(&*self.backend, insert, delete)
+	}
+	/// Query auxiliary data from key-value store.
+	fn get_aux(&self, key: &[u8]) -> error::Result<Option<Vec<u8>>> {
+		::backend::AuxStore::get_aux(&*self.backend, key)
+	}
+}
 #[cfg(test)]
 pub(crate) mod tests {
 	use std::collections::HashMap;
