+++ core/consensus/aura/Cargo.toml
@@ -2,24 +2,26 @@
 name = "substrate-consensus-aura"
 version = "0.1.0"
 authors = ["Parity Technologies <admin@parity.io>"]
-description = "Rhododendron Round-Based consensus-algorithm for substrate"
+description = "Aura consensus algorithm for substrate"
 
 [dependencies]
-futures = "0.1.17"
-parity-codec = { version = "2.1" }
-substrate-consensus-common = { path = "../common" }
-substrate-client = { path = "../../client" }
-substrate-primitives = { path = "../../primitives" }
-substrate-network = { path = "../../network" }
-srml-support = { path = "../../../srml/support" }
-sr-primitives = { path = "../../sr-primitives" }
-sr-version = { path = "../../sr-version" }
-sr-io = { path = "../../sr-io" }
-srml-consensus = { path = "../../../srml/consensus" }
-tokio = "0.1.7"
-parking_lot = "0.4"
-error-chain = "0.12"
-log = "0.3"
+parity-codec = { version = "2.1", default-features = false }
+substrate-client = { path = "../../client", default-features = false }
+substrate-primitives = { path = "../../primitives", default-features = false }
+srml-support = { path = "../../../srml/support", default-features = false }
+sr-primitives = { path = "../../sr-primitives", default-features = false }
+sr-version = { path = "../../sr-version", default-features = false }
+sr-io = { path = "../../sr-io", default-features = false }
+substrate-consensus-aura-primitives = { path = "primitives", default-features = false }
+
+srml-consensus = { path = "../../../srml/consensus", optional = true }
+futures = { version = "0.1.17", optional = true }
+tokio = { version = "0.1.7", optional = true }
+parking_lot = { version = "0.4", optional = true }
+error-chain = { version = "0.12", optional = true }
+log = { version = "0.3", optional = true }
+substrate-consensus-common = { path = "../common", optional = true }
+substrate-network = { path = "../../network", optional = true }
 
 [dev-dependencies]
 substrate-keyring = { path = "../../keyring" }
@@ -29,14 +31,27 @@ substrate-test-client = { path = "../../test-client" }
 env_logger = { version = "0.4" }
 
 [target.'cfg(test)'.dependencies]
-substrate-network = { path = "../../network", features = ["test-helpers"] }
+substrate-network = { path = "../../network", features = ["test-helpers"], optional = true }
 
 [features]
 default = ["std"]
 std = [
+	"parity-codec/std",
+	"substrate-client/std",
 	"substrate-primitives/std",
 	"srml-support/std",
 	"sr-primitives/std",
 	"sr-version/std",
+	"sr-io/std",
+	"substrate-consensus-aura-primitives/std",
+
+	"srml-consensus",
+	"futures",
+	"tokio",
+	"parking_lot",
+	"error-chain",
+	"log",
+	"substrate-consensus-common",
+	"substrate-network",
 ]
