+++ core/test-runtime/wasm/Cargo.toml
@@ -9,6 +9,7 @@ hex-literal = { version = "0.1.0", optional = true }
 parity-codec = { version = "2.1", default-features = false }
 parity-codec-derive = { version = "2.1", default-features = false }
 substrate-primitives = { path = "../../primitives", default-features = false }
+substrate-consensus-aura-primitives = { path = "../../consensus/aura/primitives", default-features = false }
 substrate-client = { path = "../../client", default-features = false }
 sr-std = { path = "../../sr-std", default-features = false }
 sr-io = { path = "../../sr-io", default-features = false }
@@ -28,7 +29,8 @@ std = [
 	"sr-version/std",
 	"substrate-primitives/std",
 	"substrate-client/std",
-	"sr-primitives/std"
+	"sr-primitives/std",
+	"substrate-consensus-aura-primitives/std",
 ]
 
 [lib]
